rename cache.js to shared.js
